Migrate to Swift 3 for all. Made it buildable on iOS 10 device.
Changes also included
- set not to ignore Podfile.lock

When you clone this project and build on your system, you need to do
this thing (based on Xcode 8.2 and iOS 10.2 device).

- Add @escaping in front of callback parameters for ActionHandler as
  defined on top of the file
- Request to use some features in Network extension API by making a
  request to Apple via
  https://developer.apple.com/contact/network-extension/.

Right now, this changes are still waiting for response back from Apple
from step 2 above. As tested, if not request and get permission from
Apple to use core feature in Network extension as needed in this
project, it will always failed similar to "permission denied".

It seems like Apple ties team account string to be able to use such
features of Network extension API.