Fix memory leak.
Fix memory leak.
The addrinfo pointer result which is created in 
if ((retval = getaddrinfo(host, service, &hints, &result)))
should be release before the function return. If not, there will be one memory leak.
It spend me two and half week to deal with this problem, and now it will just be fine.