Fix hard-code of group identifier. Update YAML in Carthage.
As per we have fixed #13 long ago, but still it's a hard-coded of group
identifier. Now it's better as it's set inside Info.plist of Potatso
main target inside PotatsoInternal/GroupIdentifier.

If the project has any more settings to be set in the future, we will
utilize this approach.

As well, update YAML from Carthage. The version is fixed via
Cartfile.resolved.