Upgrade project fully to Swift 4 support. Remove helpshift and its feedback section.
Completely removed HelpShift from codebase as this is a starting step to
remove all user's activity or app's activity logging from the app.

Feedback section feature is removed with a result from HelpShift is
removed. As well removed local Eureka directory, the project depends on
cocoapods from now.

Cocoapods that are updated for Swift 4 supported are as follows
- PSOperations
- KeychainAccess
- AsyncSwift

In short, all cocoapods are in Swift 4 support as well as the project
codebase itself.

This fixed #57.

At last, updated README.md file to reflect changes so far.